fix: keep track of match mods instead of using delta lookup
We were using mod delta as part of a key to look up the actual mod, but this
was just a hack.  It finally broke because (apparently) a float cannot transit
the whole system without possibly picking up a tiny epsilon.  So, we now do
things correctly, by actually keeping track of each match modification by
index.