chase: switch asyncore implementation to socket/generator
Drop the asyncore/asynchat network implementation for greylag-chase, in favor
of a simpler basic socket implementation (with client generator).  The new
code is quite a bit simpler, but it can only handle a single connection at a
time.  More importantly, there's no trivial way to handle a stuck client.
(Currently we may just have cron kill the chasers every couple of hours to
cover this eventuality.)

Error reporting is improved.  If a chaser dies, it will do a much better job
of sending an explanatory error to the rally program.